+++ src/main/java/io/github/ngspace/hudder/compilers/EmptyCompiler.java
@@ -1,5 +1,8 @@
 package io.github.ngspace.hudder.compilers;
 
+import io.github.ngspace.hudder.compilers.abstractions.ATextCompiler;
+import io.github.ngspace.hudder.compilers.utils.CompileException;
+import io.github.ngspace.hudder.compilers.utils.CompileResult;
 import io.github.ngspace.hudder.config.ConfigInfo;
 
 public class EmptyCompiler extends ATextCompiler {
