+++ src/main/java/io/github/ngspace/hudder/compilers/abstractions/AVarTextCompiler.java
@@ -1,10 +1,11 @@
-package io.github.ngspace.hudder.compilers;
+package io.github.ngspace.hudder.compilers.abstractions;
 
 import static io.github.ngspace.hudder.data_management.BooleanData.getBoolean;
 import static io.github.ngspace.hudder.data_management.NumberData.getNumber;
 import static io.github.ngspace.hudder.data_management.StringData.getString;
 import static io.github.ngspace.hudder.util.MathUtils.eval;
 
+import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.config.ConfigManager;
 
 public abstract class AVarTextCompiler extends ATextCompiler {
@@ -27,6 +28,16 @@ public abstract class AVarTextCompiler extends ATextCompiler {
 		return newval;
 	}
 	
+	/**
+	 * Processes a string (ex. "1+2+fps") and returns a double if it was able to process the equation, or returns the
+	 * variable with that name.
+	 * 
+	 * @deprecated Please use V2Runtime and V2Value instead of manually processing Math variables.<br>
+	 * Will be removed by version 4.0.0
+	 * @param variable
+	 * @return
+	 */
+	@Deprecated(since = "3.5.0", forRemoval = true)
 	public Object getMathVariable(String variable) {
 		String advkey = variable.trim().replace(" ", "");
 		String[] keys = advkey.split("[^\\d\\.A-z ]");
