+++ src/main/java/io/github/ngspace/hudder/compilers/utils/CompileResult.java
@@ -1,4 +1,4 @@
-package io.github.ngspace.hudder.compilers;
+package io.github.ngspace.hudder.compilers.utils;
 
 import io.github.ngspace.hudder.meta.elements.Element;
