+++ src/main/java/io/github/ngspace/hudder/compilers/v2runtime/runtime_elements/AV2RuntimeElement.java
@@ -1,6 +1,6 @@
-package io.github.ngspace.hudder.compilers.hudderv2.runtime_elements;
+package io.github.ngspace.hudder.compilers.v2runtime.runtime_elements;
 
-import io.github.ngspace.hudder.compilers.CompileException;
+import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.meta.CompileState;
 
 //TO DO Find a better name for this class
