+++ src/main/java/io/github/ngspace/hudder/compilers/v2runtime/runtime_elements/BasicConditionV2RuntimeElement.java
@@ -1,8 +1,8 @@
-package io.github.ngspace.hudder.compilers.hudderv2.runtime_elements;
+package io.github.ngspace.hudder.compilers.v2runtime.runtime_elements;
 
-import io.github.ngspace.hudder.compilers.CompileException;
-import io.github.ngspace.hudder.compilers.CompileResult;
-import io.github.ngspace.hudder.compilers.AConditionCompiler;
+import io.github.ngspace.hudder.compilers.abstractions.AConditionCompiler;
+import io.github.ngspace.hudder.compilers.utils.CompileException;
+import io.github.ngspace.hudder.compilers.utils.CompileResult;
 import io.github.ngspace.hudder.config.ConfigInfo;
 import io.github.ngspace.hudder.meta.CompileState;
