+++ src/main/java/io/github/ngspace/hudder/compilers/v2runtime/runtime_elements/WhileV2RuntimeElement.java
@@ -1,7 +1,7 @@
-package io.github.ngspace.hudder.compilers.hudderv2.runtime_elements;
+package io.github.ngspace.hudder.compilers.v2runtime.runtime_elements;
 
-import io.github.ngspace.hudder.compilers.ATextCompiler;
-import io.github.ngspace.hudder.compilers.CompileException;
+import io.github.ngspace.hudder.compilers.abstractions.ATextCompiler;
+import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.config.ConfigInfo;
 import io.github.ngspace.hudder.meta.CompileState;
