+++ src/main/java/io/github/ngspace/hudder/config/ConfigInfo.java
@@ -12,11 +12,11 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.annotations.Expose;
 
 import io.github.ngspace.hudder.Hudder;
-import io.github.ngspace.hudder.compilers.ATextCompiler;
-import io.github.ngspace.hudder.compilers.CompileException;
-import io.github.ngspace.hudder.compilers.CompileResult;
-import io.github.ngspace.hudder.compilers.Compilers;
-import io.github.ngspace.hudder.compilers.hudderv2.HudderV2Compiler;
+import io.github.ngspace.hudder.compilers.HudderV2Compiler;
+import io.github.ngspace.hudder.compilers.abstractions.ATextCompiler;
+import io.github.ngspace.hudder.compilers.utils.CompileException;
+import io.github.ngspace.hudder.compilers.utils.CompileResult;
+import io.github.ngspace.hudder.compilers.utils.Compilers;
 import io.github.ngspace.hudder.util.HudFileUtils;
 import net.minecraft.client.MinecraftClient;
