+++ src/main/java/io/github/ngspace/hudder/meta/methods/DecimalMethods.java
@@ -1,7 +1,7 @@
 package io.github.ngspace.hudder.meta.methods;
 
-import io.github.ngspace.hudder.compilers.ATextCompiler;
-import io.github.ngspace.hudder.compilers.CompileException;
+import io.github.ngspace.hudder.compilers.abstractions.ATextCompiler;
+import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.config.ConfigInfo;
 import io.github.ngspace.hudder.meta.CompileState;
 import io.github.ngspace.hudder.meta.MethodValue;
